Add support for UNIQUE(FLATMAP()) as an alternative to AGGREGATE
This change adds a simpler combination of UNIQUE and FLATMAP as an alternative to the custom AGGREGATE formulajs function.

Once we've updated card type definitions to use UNIQUE and FLATMAP and verified everything is working nicely, we can remove the AGGREGATE code.

I've also used lodash's `isMatch` function to more accurately/cleanly check for the 'special cases' of aggregating $events.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>