Make 'this' context object also available as 'contract'.
This is an intermediate step towards removing 'this' as a context
object when evaluating expressions using static-eval (as it causes
problems due to being a restricted keyword).

The next step will involve modifying the evaluate and evaluateObject
functions so that evaluate is not opinionated about naming context
variables.

In the meantime, we log a warning if the old 'this' reference is still
found in the code.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>